+++ eth_defi/anvil.py
@@ -95,6 +95,12 @@ def _launch(cmd: str, **kwargs) -> Tuple[psutil.Popen, List[str]]:
                 f"Ignoring invalid commandline setting for anvil: " f'"{key}" with value "{value}".',
                 InvalidArgumentWarning,
             )
+
+    # USDC hack
+    # Some contracts are too large to deploy when they are compiled unoptimized
+    # TODO: Move to argument
+    cmd_list += ["--code-size-limit", "99999"]
+
     final_cmd_str = " ".join(cmd_list)
     logger.info("Launching anvil: %s", final_cmd_str)
     out = DEVNULL if sys.platform == "win32" else PIPE