+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Current
+
+- Add USDC (Centre FiatToken)
+- Add EIP-712
+- Add EIP-3009
+- Add `transferWithAuthorization` and `receivedWithAuthorization`
+- Add Enzyme vault USDC payment forwarder allowing single click purchases (no `approve` step)
+- Fix: Don't try to `trace_transaction` unless we know we are on Anvil
+
 # 0.19.2
 
 - Add: Enzyme's FundValueCalculator contract as part of the deployment